requirements.txt
numpy==1.19.2
matplotlib==3.3.2
pyqt==5.9.2

